+++ weblate/wsgi.py
@@ -20,9 +20,24 @@ import os
 
 from django.core.wsgi import get_wsgi_application
 
+
+def preload_url_patterns():
+    """
+    Ensures Django URL resolver is loaded.
+
+    This avoids expensive load with a first request and makes memory sharing work
+    better between uwsgi workers.
+    """
+    from django.conf import settings
+    from django.urls import get_resolver
+
+    return get_resolver(settings.ROOT_URLCONF).url_patterns
+
+
 os.environ.setdefault("DJANGO_SETTINGS_MODULE", "weblate.settings")
 
 # This application object is used by any WSGI server configured to use this
 # file. This includes Django's development server, if the WSGI_APPLICATION
 # setting points here.
 application = get_wsgi_application()
+preload_url_patterns()